refactor: consolidate FMEA service
