+++ mainappsrc/__init__.py
@@ -38,7 +38,8 @@ _submodule_map = {
     "core.ui_setup": "ui.ui_setup",
     "node_utils": "gui.utils.node_utils",
     "core.node_utils": "gui.utils.node_utils",
-    "fmea_service": "core.fmea_service",
+    "fmea_service": "services.safety_analysis.safety_analysis_service",
+    "core.fmea_service": "services.safety_analysis.safety_analysis_service",
     "safety_analysis": "core.safety_analysis",
     "automl_core": "core.automl_core",
     # Managers
